feat- add vscode file
