--- ExampleService/ExampleService_WebApi/DataTransferObjects/TaskDTO.cs
@@ -1,31 +1,0 @@
-
-using System.ComponentModel.DataAnnotations;
-
-namespace ExampleService_WebApi;
-
-/// <summary>
-/// DTO of Task Model
-/// </summary>
-public class TaskDTO
-{
-    /// <summary>
-    /// Name of the task
-    /// </summary>
-    /// 
-    
-    public string  TaskName{ get; set; }
-
-    /// <summary>
-    /// Description about the task
-    /// </summary>
-    /// 
-   
-    public string TaskDescription{ get; set;}
-
-    /// <summary>
-    /// Status detail about the task
-    /// </summary>
-    /// 
-   
-    public string TaskStatus{ get; set;}
-}
