--- ExampleService/ExampleService_WebApi/PortLayers/InterfaceClass/TaskInterfaceClass.cs
@@ -1,21 +1,0 @@
-namespace ExampleService_WebApi.PortLayer.InterfaceClass;
-
-/// <summary>
-/// Interface class for loose coupling
-/// </summary>
-
-public interface ITaskInterface
-{
-
-    /// <summary>
-    /// returns the list of task
-    /// </summary>
-    /// <returns>task which are already added</returns>
-    List<TaskModel> ListofTask();
-
-    /// <summary>
-    /// add a new task
-    /// </summary>
-    /// <param name="task">request for task model</param>
-    void AddNewTask(TaskModel task);
-}
