+++ Todo_Application/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "dotnet.defaultSolution": "ToDo.Application.sln"
+}
