+++ Todo_Application/ToDo.Application.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33723.286
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Todo.UnitTest", "Todo.UnitTesting\Todo.UnitTest.csproj", "{3E042F49-E2D7-4A1A-8DE2-44AA3466A8A7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Todo.TaskService", "Todo.TaskServices\Todo.TaskService.csproj", "{5B03A59C-B96A-4A27-BD27-C61B23416ABA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3E042F49-E2D7-4A1A-8DE2-44AA3466A8A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3E042F49-E2D7-4A1A-8DE2-44AA3466A8A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3E042F49-E2D7-4A1A-8DE2-44AA3466A8A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3E042F49-E2D7-4A1A-8DE2-44AA3466A8A7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B03A59C-B96A-4A27-BD27-C61B23416ABA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5B03A59C-B96A-4A27-BD27-C61B23416ABA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5B03A59C-B96A-4A27-BD27-C61B23416ABA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5B03A59C-B96A-4A27-BD27-C61B23416ABA}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4A1F39A3-7C8D-4C7E-B2C1-15E6C7956DFC}
+	EndGlobalSection
+EndGlobal
