+++ Todo_Application/Todo.TaskServices/Core/CustomException/ExceptionForTaskModelRequest.cs
@@ -1,0 +1,22 @@
+﻿namespace Todo_TaskService.CoreLayers.CustomException
+
+{
+    /// <summary>
+        /// Exceptions for Invalid TaskModel request
+        /// </summary>
+
+    [Serializable]
+    public class ExceptionForTaskModelRequest: Exception
+    {
+        public ExceptionForTaskModelRequest()
+        {
+                
+        }
+        public ExceptionForTaskModelRequest(string message) 
+            : base(string.Format("Request Data : {0}",message))
+        { 
+
+        }
+
+    }
+}
